Update web page generator for Thomas's nav structure
This begins the work to transition councilmatic to the
slightly-different structure present in Thomas's designs for the app.
Primarily this is because, now, each committee has a "Past" and
"Upcoming" page.

This necessitates a difference between `/2019/city-council.html`
(which will be the "Past" link when clicked during 2019) and a new
page that includes only the upcoming events. I've named this path:
`/upcoming/city-council.html`

The "Upcoming" page replaces the sidebar logic.