JSON file datetime now extracted from file name
Datetime information for the JSON files to be compared was previously extracted using the os stat method.  But this datetime data is changed when the files are compressed and moved (a testing issue, not a production issue).  Now the datetime for the file is acquired from the 14-digit string in the file name.